fix cython typing issues by using doubles

+++ octobot_trading/data/portfolio.pxd
@@ -32,11 +32,11 @@ cdef class Portfolio(Initializable):
 
     cdef public dict portfolio
 
-    cpdef float get_currency_portfolio(self, str currency, str portfolio_type=*)
+    cpdef double get_currency_portfolio(self, str currency, str portfolio_type=*)
     cpdef void update_portfolio_available(self, Order order, bint is_new_order=*)
     cpdef void reset_portfolio_available(self, str reset_currency=*, object reset_quantity=*)
-    cpdef float get_currency_from_given_portfolio(self, str currency, str portfolio_type=*)
+    cpdef double get_currency_from_given_portfolio(self, str currency, str portfolio_type=*)
 
-    cdef void _update_portfolio_data(self, str currency, float value, bint total=*, bint available=*)
+    cdef void _update_portfolio_data(self, str currency, double value, bint total=*, bint available=*)
     cdef void _update_portfolio_available(self, Order order, bint factor=*)
     cdef bint _check_available_should_update(self, Order order)
