+++ octobot_trading/trades/trade_factory.pxd
@@ -23,18 +23,18 @@ cpdef Trade create_trade_instance_from_raw(Trader trader, dict raw_order)
 
 cpdef Trade create_trade_from_order(Order order,
                                     object close_status=*,
-                                    float canceled_time=*,
-                                    float creation_time=*,
-                                    float executed_time=*)
+                                    double canceled_time=*,
+                                    double creation_time=*,
+                                    double executed_time=*)
 
 cpdef Trade create_trade_instance(Trader trader,
                                   object order_type,
                                   str symbol,
                                   object status=*,
                                   str order_id=*,
-                                  float filled_price=*,
-                                  float quantity_filled=*,
-                                  float total_cost=*,
-                                  float canceled_time=*,
-                                  float creation_time=*,
-                                  float executed_time=*)
+                                  double filled_price=*,
+                                  double quantity_filled=*,
+                                  double total_cost=*,
+                                  double canceled_time=*,
+                                  double creation_time=*,
+                                  double executed_time=*)
